chore: move  audit directory to global
