+++ backend/src/main/java/com/cbu/backend/studygroup/StudyGroup.java
@@ -1,8 +1,8 @@
 package com.cbu.backend.studygroup;
 
-import com.cbu.backend.config.audit.AuditListener;
-import com.cbu.backend.config.audit.Auditable;
-import com.cbu.backend.global.BaseTime;
+import com.cbu.backend.global.audit.AuditListener;
+import com.cbu.backend.global.audit.Auditable;
+import com.cbu.backend.global.audit.BaseTime;
 import com.cbu.backend.member.domain.Member;
 
 import lombok.*;
